Code-recipes
A link to an example of adding a custom Ajax command to Add to Cart form variation updates. Posted by Ryan Szrama  in [Drupal Answers] on StackExchange, as an answer to a my question (https://drupal.stackexchange.com/questions/293877/commerce-hook-attributes-refresh-alter). I think It's a very good receip and It sould be reachable from here.